Add files via upload
"This GitHub repository features a clone of the renowned music streaming platform Spotify, built solely using HTML and CSS. The project aims to recreate the essential features and design elements of Spotify in a static web format. Dive into the code to explore how HTML and CSS were utilized to mimic Spotify's user interface and layout. Whether you're looking to learn web development or simply admire the design, this project serves as a valuable resource."